% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_statistics.R
\name{calculate_local_statistics}
\alias{calculate_local_statistics}
\title{Calculate local statistics}
\usage{
calculate_local_statistics(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
An object of class "local_statistics".
}
\description{
Calculates the mean, standard deviation, minimum and maximum
values of a numeric vector using the Rcpp implementation.
}
