% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_statistics.R
\name{local_statistics}
\alias{local_statistics}
\title{Create a local statistics object}
\usage{
local_statistics(mean, sd, min, max)
}
\arguments{
\item{mean}{The mean value.}

\item{sd}{The standard deviation.}

\item{min}{The minimum value.}

\item{max}{The maximum value.}
}
\value{
An object of class "local_statistics".
}
\description{
Creates an object containing the mean, standard deviation,
minimum and maximum values.
}
